ml-models/ocr-pipeline/v1/requirements.txt
flask==2.0.1
boto3==1.26.0
psycopg2-binary==2.9.5
werkzeug==2.0.3
requests==2.28.1